Update the devfile samples fetching and some css tweaks
