+++ src/components/Page/Page.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
 import { Text, TextVariants, Title } from '@patternfly/react-core';
+import classNames from 'classnames';
 import BreadCrumbs from '../../shared/components/breadcrumbs/BreadCrumbs';
 import './Page.scss';
 
@@ -7,22 +8,33 @@ type PageProps = {
   breadcrumbs?: { path: string; name: string }[];
   heading: string;
   description?: string;
+  isSection?: boolean;
   children: React.ReactNode;
 };
 
-export const Page: React.FC<PageProps> = ({ breadcrumbs, heading, description, children }) => {
+export const Page: React.FC<PageProps> = ({
+  breadcrumbs,
+  heading,
+  description,
+  children,
+  isSection,
+}) => {
   return (
     <div className="hacDev-page">
-      {breadcrumbs && <BreadCrumbs className="hacDev-page__breadcrumb" breadcrumbs={breadcrumbs} />}
-      <Title className="hacDev-page__heading" headingLevel="h1" size="4xl">
-        {heading}
-      </Title>
-      {description && (
-        <Text className="hacDev-page__description" component={TextVariants.p}>
-          {description}
-        </Text>
-      )}
-      {children}
+      <div className="hacDev-page__section">
+        {breadcrumbs && (
+          <BreadCrumbs className="hacDev-page__breadcrumb" breadcrumbs={breadcrumbs} />
+        )}
+        <Title className="hacDev-page__heading" headingLevel="h1" size="4xl">
+          {heading}
+        </Title>
+        {description && (
+          <Text className="hacDev-page__description" component={TextVariants.p}>
+            {description}
+          </Text>
+        )}
+      </div>
+      <div className={classNames({ 'hacDev-page__section': isSection })}>{children}</div>
     </div>
   );
 };
