--- src/components/ReviewComponents/ReviewComponentsForm.scss
@@ -1,7 +1,0 @@
-.hacDev-review-form {
-
-  &__components-section {
-    width: 100vw;
-    transform: translateX(calc(0px - var(--pf-global--spacer--lg)));
-  }
-}
